pyproject.toml
[project]
name = "lamdba-update-lead-status"
version = "0.1.0"
description = "Lambda to update lead status on Notion"
authors = [
    {name = "LeonardoLeiteMeira",email = "leonardo.leitemeira10@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "notion-client (>=2.3.0,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
